refactor(fs): split commands/fs.rs into security/listing/watch/io/trash/open
The 1017-LOC commands/fs.rs covered six unrelated concerns (path
security, listing/walking, watch lifecycle, read/write/create/rename,
trash with home-trash fallback, OS-level reveal/open). Split it into a
directory + mod.rs, mirroring the existing git/ and shell/ layouts.

Pure relocation — no logic changes. The atomic-write block in
fs_write_file stays inline (single-use, tightly cohesive). Path
validation helpers live in security.rs as pub(super) so every sibling
submodule consumes them through a single trust boundary, matching the
"trust boundary in one place" framing the original module doc stated.

Public surface preserved: mod.rs glob re-exports each submodule so
`commands::fs::fs_*` continues to resolve in lib.rs. Glob (rather than
named) re-export is required so Tauri's macro-generated `__cmd__*`
shadows stay reachable to `generate_handler!` — same pattern as
shell/mod.rs.

Tests migrated alongside the code they exercise: listing tests to
listing.rs, write/read/rename tests to io.rs, trash-naming test to
trash.rs. All 8 original cases pass under their new module paths.

Verified: cargo check + clippy -D warnings + cargo test --lib
(158 passing) + nix develop -c check (full devshell gate: clippy + tsc
+ ESLint + cargo test + vitest, 1124 vitest cases passing) + treefmt.
No dev build was running, so the optional aethon-debug UI smoke test
was skipped — covered by the unit tests since this is pure module
relocation.
